// apps/dashboard/src/payload-generated-schema.ts
export const enum_transactions_adapter = pgEnum('enum_transactions_adapter', ['evm', 'solana', 'starknet']);
export const enum_transactions_tracker = pgEnum('enum_transactions_tracker', [
  'ethereum',
  'safe',
  'gelato',
  'solana',
  'erc4337',
]);
export const enum_transactions_status = pgEnum('enum_transactions_status', ['Failed', 'Success', 'Replaced']);
export const enum_transactions_sync_status = pgEnum('enum_transactions_sync_status', ['synced', 'pending-sync']);
export const enum_transactions_aml_status = pgEnum('enum_transactions_aml_status', [
  'not_applicable',
  'pending',
  'passed',
  'flagged',
  'failed',
]);
export const enum_apps_environment = pgEnum('enum_apps_environment', ['live', 'test']);
export const enum_apps_kind = pgEnum('enum_apps_kind', ['basic', 'payments']);
// ...
export const enum_app_invoices_status = pgEnum('enum_app_invoices_status', [
  'draft',
  'pending',
  'paid',
  'failed',
  'expired',
  'cancelled',
]);
export const enum_app_invoices_currency = pgEnum('enum_app_invoices_currency', ['USD']);
// ...
export const transactions = pgTable(
  'transactions',
  {
    id: serial('id').primaryKey(),
    owner: varchar('owner_id')
      .notNull()
      .references(() => organizations.id, {
        onDelete: 'set null',
      }),
    app: varchar('app_id').references(() => apps.id, {
      onDelete: 'set null',
    }),
    appName: varchar('app_name').notNull().default('Web3 App'),
    txKey: varchar('tx_key').notNull(),
    chainId: varchar('chain_id').notNull(),
    from: varchar('from').notNull(),
    type: varchar('type').notNull(),
    connectorType: varchar('connector_type').notNull(),
    adapter: enum_transactions_adapter('adapter').notNull(),
    tracker: enum_transactions_tracker('tracker').notNull(),
    status: enum_transactions_status('status'),
    syncStatus: enum_transactions_sync_status('sync_status').default('synced'),
    pending: boolean('pending').notNull().default(true),
    isError: boolean('is_error').default(false),
    localTimestamp: numeric('local_timestamp', { mode: 'number' }).notNull(),
    finishedTimestamp: numeric('finished_timestamp', { mode: 'number' }),
    title: jsonb('title'),
    description: jsonb('description'),
    error: jsonb('error'),
    payload: jsonb('payload'),
    confirmations: varchar('confirmations'),
    requiredConfirmations: numeric('required_confirmations', { mode: 'number' }),
    rpcUrl: varchar('rpc_url'),
    appInvoiceId: varchar('app_invoice_id_id').references(() => app_invoices.id, {
      onDelete: 'set null',
    }),
    amlStatus: enum_transactions_aml_status('aml_status').default('not_applicable'),
    amlRiskScore: numeric('aml_risk_score', { mode: 'number' }).default(0),
    amlProviderData: jsonb('aml_provider_data'),
    hash: varchar('hash'),
    to: varchar('to'),
    value: varchar('value'),
    input: varchar('input'),
    nonce: numeric('nonce', { mode: 'number' }),
    maxFeePerGas: varchar('max_fee_per_gas'),
    maxPriorityFeePerGas: varchar('max_priority_fee_per_gas'),
    replacedTxHash: varchar('replaced_tx_hash'),
    bundlerUrl: varchar('bundler_url'),
    pimlicoApiKey: varchar('pimlico_api_key'),
    fee: numeric('fee', { mode: 'number' }),
    slot: numeric('slot', { mode: 'number' }),
    recentBlockhash: varchar('recent_blockhash'),
    instructions: jsonb('instructions'),
    actualFee: jsonb('actual_fee'),
    contractAddress: varchar('contract_address'),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 }).defaultNow().notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 }).defaultNow().notNull(),
  },
  (columns) => [
    index('transactions_owner_idx').on(columns.owner),
    index('transactions_app_idx').on(columns.app),
    index('transactions_app_name_idx').on(columns.appName),
    uniqueIndex('transactions_tx_key_idx').on(columns.txKey),
    index('transactions_chain_id_idx').on(columns.chainId),
    index('transactions_from_idx').on(columns.from),
    index('transactions_adapter_idx').on(columns.adapter),
    index('transactions_status_idx').on(columns.status),
    index('transactions_sync_status_idx').on(columns.syncStatus),
    index('transactions_local_timestamp_idx').on(columns.localTimestamp),
    index('transactions_app_invoice_id_idx').on(columns.appInvoiceId),
    index('transactions_aml_status_idx').on(columns.amlStatus),
    index('transactions_hash_idx').on(columns.hash),
    index('transactions_to_idx').on(columns.to),
    index('transactions_updated_at_idx').on(columns.updatedAt),
    index('transactions_created_at_idx').on(columns.createdAt),
  ],
);
// ...
export const apps = pgTable(
  'apps',
  {
    id: varchar('id').primaryKey(),
    isActive: boolean('is_active').default(true),
    name: varchar('name').notNull(),
    organization: varchar('organization_id')
      .notNull()
      .references(() => organizations.id, {
        onDelete: 'set null',
      }),
    environment: enum_apps_environment('environment').notNull().default('test'),
    kind: enum_apps_kind('kind').notNull().default('basic'),
    publicKey: varchar('public_key'),
    secretKey: varchar('secret_key'),
    secretKeyHash: varchar('secret_key_hash'),
    alchemyApiKey: varchar('alchemy_api_key'),
    quickNodeApiKey: varchar('quick_node_api_key'),
    quickNodeAppName: varchar('quick_node_app_name'),
    gelatoApiKey: varchar('gelato_api_key'),
    pimlicoApiKey: varchar('pimlico_api_key'),
    paymentSettings_amlEnabled: boolean('payment_settings_aml_enabled').default(false),
    paymentSettings_goPlusApiKey: varchar('payment_settings_go_plus_api_key'),
    paymentSettings_goPlusApiSecret: varchar('payment_settings_go_plus_api_secret'),
    paymentSettings_appInvoiceTemplate: jsonb('payment_settings_app_invoice_template'),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 }).defaultNow().notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 }).defaultNow().notNull(),
  },
  (columns) => [
    index('apps_is_active_idx').on(columns.isActive),
    index('apps_organization_idx').on(columns.organization),
    index('apps_environment_idx').on(columns.environment),
    index('apps_kind_idx').on(columns.kind),
    uniqueIndex('apps_public_key_idx').on(columns.publicKey),
    uniqueIndex('apps_secret_key_idx').on(columns.secretKey),
    index('apps_secret_key_hash_idx').on(columns.secretKeyHash),
    index('apps_updated_at_idx').on(columns.updatedAt),
    index('apps_created_at_idx').on(columns.createdAt),
  ],
);
// ...
export const organizations = pgTable(
  'organizations',
  {
    id: varchar('id').primaryKey(),
    name: varchar('name').notNull(),
    slug: varchar('slug'),
    createdBy: varchar('created_by').notNull(),
    quotaBalance: numeric('quota_balance', { mode: 'number' }).notNull().default(100000000),
    rpsLimit: numeric('rps_limit', { mode: 'number' }).notNull().default(100000),
    rpsPaidLimit: numeric('rps_paid_limit', { mode: 'number' }).notNull().default(100000),
    rpsExpiresAt: timestamp('rps_expires_at', { mode: 'string', withTimezone: true, precision: 3 }),
    quotaUsed: numeric('quota_used', { mode: 'number' }).notNull().default(0),
    rpsForever: boolean('rps_forever').default(true),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 }).defaultNow().notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 }).defaultNow().notNull(),
  },
  (columns) => [
    uniqueIndex('organizations_slug_idx').on(columns.slug),
    index('organizations_created_by_idx').on(columns.createdBy),
    index('organizations_updated_at_idx').on(columns.updatedAt),
    index('organizations_created_at_idx').on(columns.createdAt),
  ],
);
// ...
export const app_accepted_payments = pgTable(
  'app_accepted_payments',
  {
    id: varchar('id').primaryKey(),
    app: varchar('app_id')
      .notNull()
      .references(() => apps.id, {
        onDelete: 'set null',
      }),
    organization: varchar('organization_id')
      .notNull()
      .references(() => organizations.id, {
        onDelete: 'set null',
      }),
    name: varchar('name').notNull(),
    walletAddressToReceivePayment: varchar('wallet_address_to_receive_payment').notNull(),
    tokenAddress: varchar('token_address').notNull(),
    chainId: numeric('chain_id', { mode: 'number' }).notNull(),
    symbol: varchar('symbol').notNull(),
    decimals: numeric('decimals', { mode: 'number' }).notNull().default(6),
    markup: numeric('markup', { mode: 'number' }).default(0),
    discount: numeric('discount', { mode: 'number' }).default(0),
    priceFeedAddress: varchar('price_feed_address').notNull(),
    isActive: boolean('is_active').default(true),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 }).defaultNow().notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 }).defaultNow().notNull(),
  },
  (columns) => [
    index('app_accepted_payments_app_idx').on(columns.app),
    index('app_accepted_payments_organization_idx').on(columns.organization),
    index('app_accepted_payments_updated_at_idx').on(columns.updatedAt),
    index('app_accepted_payments_created_at_idx').on(columns.createdAt),
  ],
);

export const app_invoices = pgTable(
  'app_invoices',
  {
    id: varchar('id').primaryKey(),
    app: varchar('app_id')
      .notNull()
      .references(() => apps.id, {
        onDelete: 'set null',
      }),
    organization: varchar('organization_id')
      .notNull()
      .references(() => organizations.id, {
        onDelete: 'set null',
      }),
    status: enum_app_invoices_status('status').notNull().default('draft'),
    baseAmount: numeric('base_amount', { mode: 'number' }).notNull(),
    totalAmount: numeric('total_amount', { mode: 'number' }).notNull(),
    currency: enum_app_invoices_currency('currency').notNull().default('USD'),
    cryptoAmountExpected: varchar('crypto_amount_expected').notNull(),
    acceptedPayment: varchar('accepted_payment_id')
      .notNull()
      .references(() => app_accepted_payments.id, {
        onDelete: 'set null',
      }),
    txHash: varchar('tx_hash'),
    originWallet: varchar('origin_wallet'),
    metadata: jsonb('metadata'),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 }).defaultNow().notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 }).defaultNow().notNull(),
  },
  (columns) => [
    index('app_invoices_app_idx').on(columns.app),
    index('app_invoices_organization_idx').on(columns.organization),
    index('app_invoices_status_idx').on(columns.status),
    index('app_invoices_accepted_payment_idx').on(columns.acceptedPayment),
    uniqueIndex('app_invoices_tx_hash_idx').on(columns.txHash),
    index('app_invoices_updated_at_idx').on(columns.updatedAt),
    index('app_invoices_created_at_idx').on(columns.createdAt),
  ],
);